int=> int16_t , maintenant ça boote!

+++ Sources/globals.h
@@ -88,8 +88,8 @@ uint8_t pb_aquiz1 = 0, pb_aquiz2 = 0;
 uint16_t old_btn[3] = {0};
 
 // camera analyse_2
-int bandes_1[128][2]={0}; // bandes (0,1 et leur largeur)
-int bandes_2[128][2]={0}; // bandes (0,1 et leur largeur)
-int nb_bandes_1=0;
-int nb_bandes_2=0;
-int ligne_arrivee=0;
+uint8_t bandes_1[128][2]={0}; // bandes (0,1 et leur largeur)
+uint8_t bandes_2[128][2]={0}; // bandes (0,1 et leur largeur)
+uint8_t nb_bandes_1=0;
+uint8_t nb_bandes_2=0;
+uint8_t ligne_arrivee=0;
